// frontend/src/api/models/analysisRuleView.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { AnalysisRuleType } from './analysisRuleType'
import type { AnalysisRuleViewExcludeScope } from './analysisRuleViewExcludeScope'
import type { AnalysisRuleViewInflowScope } from './analysisRuleViewInflowScope'
import type { AnalysisRuleViewOutflowScope } from './analysisRuleViewOutflowScope'

export interface AnalysisRuleView {
  id: string
  name: string
  type: AnalysisRuleType
  excludeScope: AnalysisRuleViewExcludeScope
  inflowScope: AnalysisRuleViewInflowScope
  outflowScope: AnalysisRuleViewOutflowScope
  /** @nullable */
  archivedAt?: string | null
  createdAt: string
  updatedAt: string
}